Refactor MoveInsn class
